% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configs.R
\name{configs}
\alias{configs}
\alias{config_plots}
\alias{config_estimates}
\alias{config_trends}
\alias{config_agegroups}
\title{Load config files}
\usage{
config_plots()

config_estimates()

config_trends()

config_agegroups()
}
\description{
These variables store the data found in the config files. This config data
helps build the server and ui of the Tabby application
}
