% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/description-panel.R
\name{descriptionPanel}
\alias{descriptionPanel}
\title{Construct a Description Panel from Content intended for a Navbar Tab}
\usage{
descriptionPanel(tabclass, content)
}
\description{
The tabclass will be appended with " tab-pane", so for example to construct
the about-tab which is the first to display to the user we would use
tabclass = "about-tab active".
}
