% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/description-panel.R
\name{emptyPanel}
\alias{emptyPanel}
\title{Construct an Empty Panel where Content Used to Be}
\usage{
emptyPanel(tabclass)
}
\description{
This function is used in the ui.R to replace tab-content with
empty content so that when the user clicks on different tabs
the content is swapped out properly.
}
