% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-data.R
\name{model-results}
\alias{model-results}
\alias{results_estimates}
\alias{results_trends}
\alias{results_agegroups}
\alias{data_estimates}
\alias{data_trends}
\alias{data_agegroups}
\title{Load Raw or Transformed TB Model Results Data}
\usage{
results_estimates()

results_trends()

results_agegroups()

data_estimates()

data_trends()

data_agegroups()
}
\description{
Melted arrays, now data frames, containing the tab results of the TB model
output.
}
\details{
The `results_*` functions return the raw data, where as `data_*` functions
return formatted results ready for use in the tabby application.
}
\section{Outcomes}{


1. Incident M. TB Infection (per Mil)
2. LTBI Prevalence (%)
3. New TB Cases (per Mil)
4. MDR-TB in New TB Cases (%)
5. TB-Related Deaths (per Mil)
}

\section{Scenarios}{


1. Base case
2. Intervention scenario 1 -- TLTBI for new immigrants
3. Intervention scenario 2 -- Improved TLTBI in US
4. Intervention scenario 3 -- Better case detection
5. Intervention scenario 4 -- Better TB treatment
6. Intervention scenario 5 -- All improvements together
7. Sensitivity analysis  1 -- no transmission within the United States after 2016
8. Sensitivity analysis  2 -- no LTBI among all immigrants after 2016
}

\section{Populations}{


1. All
2. US-born
3. Foreign-born
}

\keyword{internal}
