% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\docType{package}
\name{utilities}
\alias{utilities}
\alias{utilities-package}
\alias{utilities-package}
\title{utilities: Tabby App Utilities}
\description{
Provide utilities for the Tabby TB application. These utilities
include UI building functions, data loading, config reading, and general
util functions.
}
\author{
\strong{Maintainer}: Nathan Teetor \email{nteetor@hsph.harvard.edu}

}
